# docs/ncc-api-swagger.json
{
  "swagger": "2.0",
  "info": {
    "title": "NCC - Client API",
    "description": "The NEM Community Client (NCC) provides an RESTful API to the wallet functionality. Thus enabling alternative front-end implementations be provided by the community.",
    "version": "0.6.37-BETA"
  },
  "host": "localhost:8989",
  "schemes": [
    "http"
  ],
  "basePath": "/ncc/api",
  "produces": [
    "application/json"
  ],
  "paths": {
...
    "/wallet/create": {
      "post": {
        "description": "Creates a new wallet with one account in it. The wallet is stored encrypted in the configured wallet directory. The name of the wallet file is the encoded (URL encoded) version of the wallet name. The provided password is used to encrypt the wallet.",
        "tags": [
          "Wallet"
        ],
        "consumes": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "wallet",
            "in": "body",
            "description": "The wallet name and password to be created.",
            "required": true,
            "schema": {
              "$ref": "#/definitions/WalletNamePasswordPair"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "The content of the wallet created.",
            "schema": {
              "$ref": "#/definitions/WalletViewModel"
            }
          },
          "400": {
            "description": "A bad request was sent.",
            "schema": {
              "$ref": "#/responses/BadRequest"
            }
          },
          "default": {
            "description": "Any other exception occured.",
            "schema": {
              "$ref": "#/responses/Error"
            }
          }
        }
      }
    },
    "/wallet/open": {
      "post": {
        "description": "Open a wallet. The wallet is defined by the wallet name. The wallet has to be located in the wallet storage location. The password must match the password of the given wallet. For opening a wallet at a different location the function importWallet has to be used.",
        "tags": [
          "Wallet"
        ],
        "consumes": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "wallet",
            "in": "body",
            "description": "The wallet name and password to be opened.",
            "required": true,
            "schema": {
              "$ref": "#/definitions/WalletNamePasswordPair"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "The content of the wallet opened.",
            "schema": {
              "$ref": "#/definitions/WalletViewModel"
            }
          },
          "400": {
            "description": "A bad request was sent.",
            "schema": {
              "$ref": "#/responses/BadRequest"
            }
          },
          "default": {
            "description": "Any other exception occured.",
            "schema": {
              "$ref": "#/responses/Error"
            }
          }
        }
      }
    },
    "/wallet/info": {
      "post": {
        "description": "Returns information about a wallet that is already open.",
        "tags": [
          "Wallet"
        ],
        "consumes": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "wallet",
            "in": "body",
            "description": "The wallet name.",
            "required": true,
            "schema": {
              "$ref": "#/definitions/WalletDescriptor"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "The content of the wallet.",
            "schema": {
              "$ref": "#/definitions/WalletViewModel"
            }
          },
          "400": {
            "description": "A bad request was sent.",
            "schema": {
              "$ref": "#/responses/BadRequest"
            }
          },
          "default": {
            "description": "Any other exception occured.",
            "schema": {
              "$ref": "#/responses/Error"
            }
          }
        }
      }
    },
    "/wallet/export": {
      "post": {
        "description": "Exports a wallet as a zip file.",
        "tags": [
          "Wallet"
        ],
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/zip"
        ],
        "parameters": [
          {
            "name": "wallet",
            "in": "body",
            "description": "The wallet name.",
            "required": true,
            "schema": {
              "$ref": "#/definitions/WalletDescriptor"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "A ZIP file containing the wallet and the address book.",
            "schema": {
              "type": "string"
            }
          },
          "400": {
            "description": "A bad request was sent.",
            "schema": {
              "$ref": "#/responses/BadRequest"
            }
          },
          "default": {
            "description": "Any other exception occured.",
            "schema": {
              "$ref": "#/responses/Error"
            }
          }
        }
      }
    },
    "/wallet/close": {
      "post": {
        "description": "Closes a wallet by removing it from the list of opened wallets.",
        "tags": [
          "Wallet"
        ],
        "consumes": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "wallet",
            "in": "body",
            "description": "The wallet name.",
            "required": true,
            "schema": {
              "$ref": "#/definitions/WalletDescriptor"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "The wallet is closed."
          },
          "400": {
            "description": "A bad request was sent.",
            "schema": {
              "$ref": "#/responses/BadRequest"
            }
          },
          "default": {
            "description": "Any other exception occured.",
            "schema": {
              "$ref": "#/responses/Error"
            }
          }
        }
      }
    },
    "/wallet/password/change": {
      "post": {
        "description": "Changes the password of a wallet.",
        "tags": [
          "Wallet"
        ],
        "consumes": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "wallet",
            "in": "body",
            "description": "The wallet name, password and the new password. The attribute \"newPassword\" is required.",
            "required": true,
            "schema": {
              "$ref": "#/definitions/WalletNamePasswordBag"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "The password is changed."
          },
          "400": {
            "description": "A bad request was sent.",
            "schema": {
              "$ref": "#/responses/BadRequest"
            }
          },
          "default": {
            "description": "Any other exception occured.",
            "schema": {
              "$ref": "#/responses/Error"
            }
          }
        }
      }
    },
    "/wallet/name/change": {
      "post": {
        "description": "Changes the name of a wallet.",
        "tags": [
          "Wallet"
        ],
        "consumes": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "wallet",
            "in": "body",
            "description": "The wallet name, password and the new name. The attribute \"newName\" is required.",
            "required": true,
            "schema": {
              "$ref": "#/definitions/WalletNamePasswordBag"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "The wallet name is changed."
          },
          "400": {
            "description": "A bad request was sent.",
            "schema": {
              "$ref": "#/responses/BadRequest"
            }
          },
          "default": {
            "description": "Any other exception occured.",
            "schema": {
              "$ref": "#/responses/Error"
            }
          }
        }
      }
    },
...
  },
  "definitions": {
...
    "ErrorResponseHttp": {
      "properties": {
        "timeStamp": {
          "type": "integer",
          "description": "Point in time when error happened. Time is based on NEM epoch time."
        },
        "error": {
          "type": "string",
          "description": "Error message."
        },
        "message": {
          "type": "string",
          "description": "Additional message about the error."
        },
        "status": {
          "$ref": "#/definitions/HttpErrorCodes",
          "description": "Status of the error."
        }
      },
      "example": {
        "timeStamp": 4978518,
        "error": "Not Found",
        "message": null,
        "status": 404
      }
    },
    "ErrorResponseNcc": {
      "properties": {
        "timeStamp": {
          "type": "integer",
          "description": "Point in time when error happened. Time is based on NEM epoch time."
        },
        "error": {
          "type": "string",
          "description": "Error message."
        },
        "message": {
          "type": "string",
          "description": "Additional message about the error."
        },
        "status": {
          "$ref": "#/definitions/NccErrorCodes",
          "description": "Status of the error."
        }
      },
      "example": {
        "timeStamp": 4978518,
        "error": "Not Found",
        "message": null,
        "status": 404
      }
    },
    "NccErrorCodes": {
      "type": "integer",
      "format": "int32",
      "description": "Error codes that indicate the reason for failing:  \n* 106 - WALLET_IS_NOT_OPEN, The specified wallet is not open.\n* 126 - ADDRESS_BOOK_IS_NOT_OPEN, The specified address book is not open.\n* 202 - NO_PUBLIC_KEY, The recipient public key is unknown.\n* 203 - COSIGNATORY_NO_PUBLIC_KEY, The cosignatory's public key is unknown.\n* 901 - ACCOUNT_CACHE_ERROR, There was a problem with the account cache.\n* 1000 - PRIVATE_KEY_PUBLIC_KEY_MISMATCH, The public key derived from the private key does not match the given public key.\n* 1001 - PUBLIC_KEY_ADDRESS_MISMATCH, The address derived from the public key does not match the given address.\n* 1002 - NOT_MAIN_NETWORK_ADDRESS, The address does not belong to the main network.\n* 305 - NIS_NOT_AVAILABLE, NCC is disconnected from the network.",
      "enum": [
        106,
        126,
        202,
        203,
        305,
        901,
        1000,
        1001,
        1002
      ]
    },
    "HttpErrorCodes": {
      "type": "integer",
      "format": "int32",
      "description": "Error codes that indicate the reason for failing:  \n* 400 - Bad request, e.g., mandatory parameter is missing or value is not correct.\n* 401 - Unauthorized, Request must be authorized.\n* 404 - Not Found, requested resource not found.",
      "enum": [
        400,
        401,
        404,
        500
      ]
    },
...
    "WalletDescriptor": {
      "properties": {
        "name": {
          "type": "string",
          "description": "The name of the wallet."
        }
      },
      "description": "The name of the wallet."
    },
    "WalletNamePasswordPair": {
      "properties": {
        "wallet": {
          "type": "string",
          "description": "The name of the wallet."
        },
        "password": {
          "type": "string",
          "description": "The password used for encrypting the wallet."
        }
      },
      "description": "The name and password of a wallet."
    },
    "WalletNamePasswordBag": {
      "allOf": [
        {
          "$ref": "#/definitions/WalletNamePasswordPair"
        },
        {
          "properties": {
            "newName": {
              "type": "string",
              "description": "The new name for the wallet."
            },
            "newPassword": {
              "type": "string",
              "description": "The new password for the wallet."
            },
            "account": {
              "type": "string",
              "description": "The address of an account."
            },
            "label": {
              "type": "string",
              "description": "The private label of an account."
            },
            "accountKey": {
              "type": "string",
              "description": "The private key of an account."
            },
            "token": {
              "type": "string",
              "description": "The token if it was specified."
            }
          }
        }
      ],
      "description": "A bag containing additional attributes."
    },
    "WalletViewModel": {
      "properties": {
        "wallet": {
          "type": "string",
          "description": "The name of the wallet."
        },
        "lastRefresh": {
          "type": "integer",
          "format": "int32",
          "description": "The time of the last refresh."
        },
        "primaryAccount": {
          "$ref": "#/definitions/AccountViewModel",
          "description": "The primary account of the wallet."
        },
        "otherAccounts": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/AccountViewModel"
          },
          "description": "The primary account of the wallet."
        }
      },
      "description": "The content of the wallet."
    },
    "AccountViewModel": {
      "properties": {
        "address": {
          "type": "string",
          "description": "The address of the account."
        },
        "remoteStatus": {
          "type": "string",
          "description": "The remote status of the account.\n* REMOTE - The account is remote one, and therefore RemoteStatus is not applicable for it.\n* REMOTE - The account has activated remote harvesting (but not yet active).\n* ACTIVE - The account has activated remote harvesting and remote harvesting is active.\n* DEACTIVATING - The account has deactivated remote harvesting (but remote harvesting is still active).\n* INACTIVE - The account has inactive remote harvesting, or it has deactivated remote harvesting and deactivation is operational."
        },
        "publicKey": {
          "type": "string",
          "description": "The public key of the account."
        },
        "balance": {
          "type": "integer",
          "format": "int64",
          "description": "The amount of NEMs (in uNEM) of the account."
        },
        "vestedBalance": {
          "type": "integer",
          "format": "int64",
          "description": "The vested balance of the account (in uNEM: 1.000.000uNEM = 1 NEM)."
        },
        "importance": {
          "type": "number",
          "format": "double",
          "description": "The importance of the account to the NEM network."
        },
        "harvestedBlocks": {
          "type": "integer",
          "format": "int64",
          "description": "The number of harvested blocks."
        },
        "status": {
          "type": "string",
          "description": "The status of the account.\n*UNKNOWN - The account status is unknown (needed for ncc).\n*LOCKED - The account is locked.\n*UNLOCKED - The account is unlocked"
        },
        "multisigAccounts": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/AccountInfo"
          },
          "description": "list of accounts (info) that this account is cosignatory of."
        },
        "cosignatories": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/AccountInfo"
          },
          "description": "List of accounts (info) that are cosignatories of this account."
        },
        "remotePublicKey": {
          "type": "string",
          "description": "The account's remote public key (optional)."
        },
        "remoteAddress": {
          "type": "string",
          "description": "The account's remote address (optional)."
        }
      },
      "description": "The content of the wallet."
    },
    "AccountInfo": {
      "properties": {
        "address": {
          "type": "string",
          "description": "The address of the account."
        },
        "publicKey": {
          "type": "string",
          "description": "The public key of the account."
        },
        "balance": {
          "type": "integer",
          "format": "int64",
          "description": "The amount of NEMs (in uNEM) of the account."
        },
        "vestedBalance": {
          "type": "integer",
          "format": "int64",
          "description": "The vested balance of the account (in uNEM: 1.000.000uNEM = 1 NEM)."
        },
        "importance": {
          "type": "number",
          "format": "double",
          "description": "The importance of the account to the NEM network."
        },
        "harvestedBlocks": {
          "type": "integer",
          "format": "int64",
          "description": "The number of harvested blocks."
        },
        "label": {
          "type": "string",
          "description": "The label of the account (public label)."
        }
      },
      "description": "list of accounts (info) that this account is cosignatory of."
    },
...
  },
  "responses": {
    "BadRequest": {
      "description": "Bad request, further information within error response object.",
      "schema": {
        "$ref": "#/definitions/ErrorResponseNcc"
      }
    },
    "Error": {
      "description": "Unexpected error occured.",
      "schema": {
        "$ref": "#/definitions/ErrorResponseHttp"
      }
    }
  }
}